Adding granularity parameter
git-svn-id: https://svn.ebi.ac.uk/steinbeck-research/CheMet/trunk@578 47e25a5e-c6eb-4c94-8020-4a4a81ee2334

+++ CheMet/CheMet-Visualisation/src/main/java/uk/ac/ebi/chemet/visualisation/ConsensusScorer.java
@@ -33,8 +33,8 @@ public interface ConsensusScorer {
 
     /**
      * Scores the character from a consensus alignment string. For example in blast ' ' is a mismatch, 'D' etc is
-     * a match and a '+' is a matching group
+     * a match and a '+' is a matching group. 0 = no match, 5 = some match and 10 = all match
      */
-    public float score(char match);
+    public int score(char match);
     
 }
